--- src/main/java/com/example/humorie/admin/dto/ConsultCompletionDto.java
@@ -1,12 +1,0 @@
-package com.example.humorie.admin.dto;
-
-import lombok.Getter;
-
-@Getter
-public class ConsultCompletionDto {
-
-    private String counselorName;
-
-    private String counselDate;
-
-}
